+++ sentry/src/main/java/io/sentry/EnvelopeSender.java
@@ -4,7 +4,6 @@ import io.sentry.cache.EnvelopeCache;
 import io.sentry.hints.Flushable;
 import io.sentry.hints.Retryable;
 import io.sentry.util.HintUtils;
-import io.sentry.util.LogUtils;
 import io.sentry.util.Objects;
 import java.io.BufferedInputStream;
 import java.io.File;
@@ -12,10 +11,8 @@ import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.Map;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 @ApiStatus.Internal
 public final class EnvelopeSender extends DirectoryProcessor implements IEnvelopeSender {
@@ -36,7 +33,7 @@ public final class EnvelopeSender extends DirectoryProcessor implements IEnvelop
   }
 
   @Override
-  protected void processFile(final @NotNull File file, final @Nullable Map<String, Object> hint) {
+  protected void processFile(final @NotNull File file, final @NotNull Hint hint) {
     if (!file.isFile()) {
       logger.log(SentryLevel.DEBUG, "'%s' is not a file.", file.getAbsolutePath());
       return;
@@ -56,8 +53,6 @@ public final class EnvelopeSender extends DirectoryProcessor implements IEnvelop
       return;
     }
 
-    Object sentrySdkHint = HintUtils.getSentrySdkHint(hint);
-
     try (final InputStream is = new BufferedInputStream(new FileInputStream(file))) {
       SentryEnvelope envelope = serializer.deserializeEnvelope(is);
       if (envelope == null) {
@@ -67,13 +62,15 @@ public final class EnvelopeSender extends DirectoryProcessor implements IEnvelop
         hub.captureEnvelope(envelope, hint);
       }
 
-      if (sentrySdkHint instanceof Flushable) {
-        if (!((Flushable) sentrySdkHint).waitFlush()) {
-          logger.log(SentryLevel.WARNING, "Timed out waiting for envelope submission.");
-        }
-      } else {
-        LogUtils.logIfNotFlushable(logger, sentrySdkHint);
-      }
+      HintUtils.runIfHasTypeLogIfNot(
+          hint,
+          Flushable.class,
+          logger,
+          (flushable) -> {
+            if (!flushable.waitFlush()) {
+              logger.log(SentryLevel.WARNING, "Timed out waiting for envelope submission.");
+            }
+          });
     } catch (FileNotFoundException e) {
       logger.log(SentryLevel.ERROR, e, "File '%s' cannot be found.", file.getAbsolutePath());
     } catch (IOException e) {
@@ -81,27 +78,31 @@ public final class EnvelopeSender extends DirectoryProcessor implements IEnvelop
     } catch (Throwable e) {
       logger.log(
           SentryLevel.ERROR, e, "Failed to capture cached envelope %s", file.getAbsolutePath());
-      if (sentrySdkHint instanceof Retryable) {
-        ((Retryable) sentrySdkHint).setRetry(false);
-        logger.log(SentryLevel.INFO, e, "File '%s' won't retry.", file.getAbsolutePath());
-      } else {
-        LogUtils.logIfNotRetryable(logger, sentrySdkHint);
-      }
+      HintUtils.runIfHasTypeLogIfNot(
+          hint,
+          Retryable.class,
+          logger,
+          (retryable) -> {
+            retryable.setRetry(false);
+            logger.log(SentryLevel.INFO, e, "File '%s' won't retry.", file.getAbsolutePath());
+          });
     } finally {
       // Unless the transport marked this to be retried, it'll be deleted.
-      if (sentrySdkHint instanceof Retryable) {
-        if (!((Retryable) sentrySdkHint).isRetry()) {
-          safeDelete(file, "after trying to capture it");
-          logger.log(SentryLevel.DEBUG, "Deleted file %s.", file.getAbsolutePath());
-        } else {
-          logger.log(
-              SentryLevel.INFO,
-              "File not deleted since retry was marked. %s.",
-              file.getAbsolutePath());
-        }
-      } else {
-        LogUtils.logIfNotRetryable(logger, sentrySdkHint);
-      }
+      HintUtils.runIfHasTypeLogIfNot(
+          hint,
+          Retryable.class,
+          logger,
+          (retryable) -> {
+            if (!retryable.isRetry()) {
+              safeDelete(file, "after trying to capture it");
+              logger.log(SentryLevel.DEBUG, "Deleted file %s.", file.getAbsolutePath());
+            } else {
+              logger.log(
+                  SentryLevel.INFO,
+                  "File not deleted since retry was marked. %s.",
+                  file.getAbsolutePath());
+            }
+          });
     }
   }
 
@@ -111,8 +112,7 @@ public final class EnvelopeSender extends DirectoryProcessor implements IEnvelop
   }
 
   @Override
-  public void processEnvelopeFile(
-      final @NotNull String path, final @Nullable Map<String, Object> hint) {
+  public void processEnvelopeFile(final @NotNull String path, final @NotNull Hint hint) {
     Objects.requireNonNull(path, "Path is required.");
 
     processFile(new File(path), hint);