+++ sentry/src/main/java/io/sentry/HubAdapter.java
@@ -5,7 +5,6 @@ import io.sentry.protocol.SentryTransaction;
 import io.sentry.protocol.User;
 import java.util.Date;
 import java.util.List;
-import java.util.Map;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -26,8 +25,7 @@ public final class HubAdapter implements IHub {
   }
 
   @Override
-  public @NotNull SentryId captureEvent(
-      @NotNull SentryEvent event, @Nullable Map<String, Object> hint) {
+  public @NotNull SentryId captureEvent(@NotNull SentryEvent event, @Nullable Hint hint) {
     return Sentry.captureEvent(event, hint);
   }
 
@@ -38,14 +36,12 @@ public final class HubAdapter implements IHub {
 
   @ApiStatus.Internal
   @Override
-  public @NotNull SentryId captureEnvelope(
-      @NotNull SentryEnvelope envelope, @Nullable Map<String, Object> hint) {
+  public @NotNull SentryId captureEnvelope(@NotNull SentryEnvelope envelope, @Nullable Hint hint) {
     return Sentry.getCurrentHub().captureEnvelope(envelope, hint);
   }
 
   @Override
-  public @NotNull SentryId captureException(
-      @NotNull Throwable throwable, @Nullable Map<String, Object> hint) {
+  public @NotNull SentryId captureException(@NotNull Throwable throwable, @Nullable Hint hint) {
     return Sentry.captureException(throwable, hint);
   }
 
@@ -70,7 +66,7 @@ public final class HubAdapter implements IHub {
   }
 
   @Override
-  public void addBreadcrumb(@NotNull Breadcrumb breadcrumb, @Nullable Map<String, Object> hint) {
+  public void addBreadcrumb(@NotNull Breadcrumb breadcrumb, @Nullable Hint hint) {
     Sentry.addBreadcrumb(breadcrumb, hint);
   }
 
@@ -163,7 +159,7 @@ public final class HubAdapter implements IHub {
   public @NotNull SentryId captureTransaction(
       @NotNull SentryTransaction transaction,
       @Nullable TraceState traceState,
-      @Nullable Map<String, Object> hint,
+      @Nullable Hint hint,
       @Nullable ProfilingTraceData profilingTraceData) {
     return Sentry.getCurrentHub()
         .captureTransaction(transaction, traceState, hint, profilingTraceData);