+++ sentry/src/main/java/io/sentry/Sentry.java
@@ -9,7 +9,6 @@ import java.io.File;
 import java.lang.reflect.InvocationTargetException;
 import java.util.Date;
 import java.util.List;
-import java.util.Map;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -285,7 +284,7 @@ public final class Sentry {
    * @return The Id (SentryId object) of the event
    */
   public static @NotNull SentryId captureEvent(
-      final @NotNull SentryEvent event, final @Nullable Map<String, Object> hint) {
+      final @NotNull SentryEvent event, final @Nullable Hint hint) {
     return getCurrentHub().captureEvent(event, hint);
   }
 
@@ -329,7 +328,7 @@ public final class Sentry {
    * @return The Id (SentryId object) of the event
    */
   public static @NotNull SentryId captureException(
-      final @NotNull Throwable throwable, final @Nullable Map<String, Object> hint) {
+      final @NotNull Throwable throwable, final @Nullable Hint hint) {
     return getCurrentHub().captureException(throwable, hint);
   }
 
@@ -349,7 +348,7 @@ public final class Sentry {
    * @param hint SDK specific but provides high level information about the origin of the event
    */
   public static void addBreadcrumb(
-      final @NotNull Breadcrumb breadcrumb, final @Nullable Map<String, Object> hint) {
+      final @NotNull Breadcrumb breadcrumb, final @Nullable Hint hint) {
     getCurrentHub().addBreadcrumb(breadcrumb, hint);
   }