+++ sentry/src/test/java/io/sentry/ScopeTest.kt
@@ -858,7 +858,7 @@ class ScopeTest {
 
     private fun eventProcessor(): EventProcessor {
         return object : EventProcessor {
-            override fun process(event: SentryEvent, hint: MutableMap<String, Any>?): SentryEvent? {
+            override fun process(event: SentryEvent, hint: Hint): SentryEvent? {
                 return event
             }
         }